refactor: Refactored scheduled jobs with Quartz instead of FluentScheduler
FluentScheduler is not supported anymore and causes threading issues. Also made Quartz jobs configureable via the appsettings.
