+++ src/Eurofurence.App.Server.Services/DependencyResolution/AutofacModule.cs
@@ -68,6 +68,7 @@ namespace Eurofurence.App.Server.Services.DependencyResolution
             builder.RegisterInstance(ArtistAlleyConfiguration.FromConfiguration(_configuration));
             builder.RegisterInstance(LassieConfiguration.FromConfiguration(_configuration));
             builder.RegisterInstance(MinIoConfiguration.FromConfiguration(_configuration));
+            builder.RegisterInstance(AnnouncementConfiguration.FromConfiguration(_configuration));
         }
 
         private void RegisterServices(ContainerBuilder builder)
