+++ src/Eurofurence.App.Server.Web/Jobs/UpdateFursuitCollectionGameParticipationJob.cs
@@ -1,15 +1,15 @@
 ﻿using Eurofurence.App.Server.Services.Abstractions.Fursuits;
-using FluentScheduler;
 using Microsoft.Extensions.Logging;
 using System;
 using System.Threading.Tasks;
+using Quartz;
 
 namespace Eurofurence.App.Server.Web.Jobs
 {
     public class UpdateFursuitCollectionGameParticipationJob : IJob
     {
-        private ICollectingGameService _collectingGameService;
-        private ILogger _logger;
+        private readonly ICollectingGameService _collectingGameService;
+        private readonly ILogger _logger;
 
         public UpdateFursuitCollectionGameParticipationJob(
             ILoggerFactory loggerFactory,
@@ -19,21 +19,18 @@ namespace Eurofurence.App.Server.Web.Jobs
             _logger = loggerFactory.CreateLogger(GetType());
         }
 
-        public void Execute()
+        public async Task Execute(IJobExecutionContext context)
         {
+            _logger.LogInformation($"Starting job {context.JobDetail.Key.Name}");
+
             try
             {
-                ExecuteAsync().Wait();
+                await _collectingGameService.UpdateFursuitParticipationAsync();
             }
             catch (Exception e)
             {
-                _logger.LogError("Job failed with exception {Message} {StackTrace}", e.Message, e.StackTrace);
+                _logger.LogError($"Job {context.JobDetail.Key.Name} failed with exception {e.Message} {e.StackTrace}");
             }
-        }
-
-        private async Task ExecuteAsync()
-        {
-            await _collectingGameService.UpdateFursuitParticipationAsync();
         }
     }
 }
