+++ src/Eurofurence.App.Server.Web/appsettings.sample.json
@@ -69,16 +69,29 @@
     }
   },
   "jobs": {
-    "updateNews": {
-      "secondsInterval": 60,
-      "source": {
-        "url": ""
-      }
+    "flushPrivateMessageNotifications": {
+      "enabled": true,
+      "secondsInterval": 1
+    },
+    "updateAnnouncements": {
+      "enabled": true,
+      "secondsInterval": 60
+    },
+    "updateFursuitCollectionGameParticipation": {
+      "enabled": true,
+      "secondsInterval": 60
+    },
+    "updateLostAndFound": {
+      "enabled": true,
+      "secondsInterval": 60
     }
   },
   "lassie": {
     "baseApiUrl": "",
     "apiKey": ""
+  },
+  "announcements": {
+    "url": ""
   },
   "minIo": {
     "endpoint": "minio:9000",
